Strip whitespace. Refactor.
