+++ pybossa/view/projects.py
@@ -59,7 +59,7 @@ from pybossa.util import (Pagination, admin_required, get_user_id_or_ip, rank,
                           check_annex_response,
                           process_annex_load, process_tp_components,
                           process_table_component, PARTIAL_ANSWER_POSITION_KEY,
-                          SavedTaskPositionEnum, delete_redis_keys,
+                          SavedTaskPositionEnum, delete_redis_keys, get_last_name,
                           PARTIAL_ANSWER_PREFIX, PARTIAL_ANSWER_KEY)
 from pybossa.auth import ensure_authorized_to
 from pybossa.cache import projects as cached_projects, ONE_DAY
@@ -3845,22 +3845,6 @@ def notify_redundancy_updates(tasks_not_updated):
                    body=body)
         mail_queue.enqueue(send_mail, email)
 
-
-def get_last_name(fullname):
-    last_name = ''
-
-    if fullname:
-        # Remove content within parentheses in name: Jane Doe (ai)
-        cleaned_name = re.sub(r'\s*\(.*?\)', '', fullname)
-        full_name_parts = cleaned_name.split(' ')
-
-        # Check if the last part is a number and use the second last if available.
-        if full_name_parts[-1].isdigit():
-            last_name = full_name_parts[-2] if len(full_name_parts) > 1 else fullname
-        else:
-            last_name = full_name_parts[-1]
-
-    return last_name
 
 @blueprint.route('/<short_name>/assign-users', methods=['GET', 'POST'])
 @login_required
